requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-okx==0.1.5
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.3
jinja2==3.1.3
